application stop

+++ deploy_scripts/application_stop.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
 #Stopping existing node servers
-echo "Stopping any existing node servers"
-pkill node
+
+echo "CHECK NODE YG SEDANG BERJALAN"
+if ps -C node > /dev/null
+then
+  echo "MEMATIKAN NODE SERVER YANG ADA"
+  pkill node
+fi
+
+echo "NODE SERVER TELAH DIMATIKAN"
